# contracts/openapi/recommendation-api.yaml
openapi: 3.1.0
info:
  title: Recommendation Platform API
  version: 0.2.0
  description: 범용 감정 분석 기반 콘텐츠·장소 추천 플랫폼의 초기 계약
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/recommendations:
    post:
      operationId: createRecommendation
      parameters:
        - name: X-Kiosk-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
      responses:
        '200':
          description: Recommendation decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResult'
        '409':
          description: No eligible candidate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
...
  schemas:
    LocalizedText:
      type: object
      additionalProperties:
        type: string
      minProperties: 1
...
    RecommendationRequest:
      type: object
      required:
        - schema_version
        - project_code
        - kiosk_id
        - session_id
        - request_id
        - emotion_code
        - stress_score
        - language
        - requested_at
      properties:
        schema_version:
          type: integer
          enum: [1, 2]
        project_code:
          type: string
        kiosk_id:
          type: string
        session_id:
          type: string
          format: uuid
        request_id:
          type: string
          format: uuid
        emotion_code:
          type: string
        stress_score:
          type: integer
          minimum: 0
          maximum: 100
        language:
          type: string
        previous_location_id:
          type: [string, 'null']
          format: uuid
        consent_status:
          type: string
          enum: [CONSENTED, DECLINED, NOT_ASKED]
          default: NOT_ASKED
          description: 개인정보 내용이 아닌 익명 통계 활용 동의 상태
        condition_code:
          type: [string, 'null']
          description: schema version 2의 12상태 코드
        journey_sense_codes:
          type: array
          maxItems: 3
          items:
            type: string
          description: 추천할 1·2·3정거장의 감각 코드 순서
        operation_context:
          $ref: '#/components/schemas/OperationContext'
        requested_at:
          type: string
          format: date-time
    RecommendationResult:
      type: object
      required:
        - schema_version
        - recommendation_id
        - request_id
        - emotion_profile
        - item
        - location
        - display
        - source
        - policy_version
        - reasons
        - created_at
      properties:
        schema_version:
          type: integer
          enum: [1, 2]
        recommendation_id:
          type: string
          format: uuid
        request_id:
          type: string
          format: uuid
        emotion_profile:
          $ref: '#/components/schemas/EmotionProfile'
        item:
          $ref: '#/components/schemas/RecommendationItem'
        location:
          $ref: '#/components/schemas/Location'
        display:
          $ref: '#/components/schemas/DisplayConfig'
        source:
          type: string
          enum: [LOCAL, REMOTE, LOCAL_FALLBACK]
        policy_version:
          type: string
        reasons:
          type: array
          items:
            type: string
        journey:
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/JourneyStop'
        created_at:
          type: string
          format: date-time
...
    EmotionProfile:
      type: object
      required: [code, name, message, color]
      properties:
        code:
          type: string
        name:
          $ref: '#/components/schemas/LocalizedText'
        message:
          $ref: '#/components/schemas/LocalizedText'
        color:
          type: string
        icon:
          type: [string, 'null']
    RecommendationItem:
      type: object
      required: [id, type, name, description, image_url]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        name:
          $ref: '#/components/schemas/LocalizedText'
        description:
          $ref: '#/components/schemas/LocalizedText'
        image_url:
          type: string
        attributes:
          type: object
          additionalProperties: true
    JourneyStop:
      type: object
      required: [order, sense_code, item, location]
      properties:
        order:
          type: integer
          minimum: 1
          maximum: 3
        sense_code:
          type: string
        item:
          $ref: '#/components/schemas/RecommendationItem'
        location:
          $ref: '#/components/schemas/Location'
    OperationContext:
      type: object
      properties:
        raining:
          type: boolean
          default: false
        companion_type:
          type: [string, 'null']
        performance_window_open:
          type: boolean
          default: false
    Location:
      type: object
      required: [id, code, name, status, marker]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          $ref: '#/components/schemas/LocalizedText'
        status:
          type: string
          enum: [NORMAL, CONGESTED, PAUSED]
        marker:
          type: object
          required: [x_percent, y_percent]
          properties:
            x_percent:
              type: number
              minimum: 0
              maximum: 100
            y_percent:
              type: number
              minimum: 0
              maximum: 100
    DisplayConfig:
      type: object
      required: [recommendation_text, display_seconds]
      properties:
        recommendation_text:
          $ref: '#/components/schemas/LocalizedText'
        display_seconds:
          type: integer
          minimum: 1
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string